move PyBytesWriter to types::bytes module
