+++ src/lib.rs
@@ -436,7 +436,6 @@ pub mod marker;
 pub mod marshal;
 #[macro_use]
 pub mod sync;
-pub mod byteswriter;
 pub mod panic;
 pub mod pybacked;
 pub mod pycell;
